% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimalModelSearch.R
\name{.draw_confusion_matrix}
\alias{.draw_confusion_matrix}
\title{Internal helper function to draw a confusion matrix visualization}
\usage{
.draw_confusion_matrix(cm)
}
\arguments{
\item{cm}{A confusion matrix object (typically from `caret::confusionMatrix`).}
}
\description{
Internal helper function to draw a confusion matrix visualization
}
\keyword{internal}
