% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_descrip.R
\name{.plot_barplot_ind}
\alias{.plot_barplot_ind}
\title{Internal helper function for plotting barplots}
\usage{
.plot_barplot_ind(x, var_name, col_idx, max_levels = 13)
}
\arguments{
\item{x}{The vector of data for the variable.}

\item{var_name}{The name of the variable (for plot title/labels).}

\item{col_idx}{The column index (used for default color selection).}

\item{max_levels}{The maximum number of unique levels before percentages are omitted (default 13).}
}
\description{
Internal helper function for plotting barplots
}
\keyword{internal}
